requirements.txt
Django==4.2.7
deepface==0.0.79
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
pandas==2.0.3
tensorflow==2.13.0
scikit-learn==1.3.0
matplotlib==3.7.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-decouple==3.8
